renamed cka base class
